fix: Clone element improvements

+++ inc/customizer/extend-custom-controls/src/builder-layout/item-component.js
@@ -7,18 +7,28 @@ const ItemComponent = props => {
 
 	const deleteItem = (props) => {
 
-		sessionStorage.setItem('forceRemoveComponent',  JSON.stringify(choices[props.item])  )
+
+		sessionStorage.setItem('forceRemoveInProgress', true);
+
+		var event = new CustomEvent('AstraBuilderDeleteSectionControls', {
+			'detail': choices[props.item]
+		});
+		document.dispatchEvent(event);
 
 		let forceRemoveSection = choices[props.item];
 		delete choices[props.item];
 
 		const component_track = wp.customize('astra-settings[cloned-component-track]').get();
 
-		let removing_index= forceRemoveSection.section.match(/\d+$/)[0];
+		let removing_index= parseInt( forceRemoveSection.section.match(/\d+$/)[0] );
 		let existing_component_count = component_track[ forceRemoveSection.builder + '-' + forceRemoveSection.type ];
-
 		let finalArray = component_track['removed-items'];
-		finalArray.push(forceRemoveSection.section);
+
+		// In removing last element.
+		if( removing_index != parseInt( AstraBuilderCustomizerData.component_limit ) ) {
+			finalArray.push(forceRemoveSection.section);
+		}
+
 		finalArray = finalArray.filter(function(el, index, arr) {
 			return index == arr.indexOf(el);
 		});
@@ -29,11 +39,11 @@ const ItemComponent = props => {
 				existing_component_count = existing_component_count - 1;
 				component_track[ forceRemoveSection.builder + '-' + forceRemoveSection.type ] = existing_component_count;
 
-				var index = finalArray.indexOf( forceRemoveSection.section.replace(/[0-9]/g, existing_component_count) );
+				var index = finalArray.indexOf( forceRemoveSection.section.replace(/[0-9]+/g, existing_component_count) );
 				if (index !== -1) {
 					finalArray.splice(index, 1);
 				} else {
-					var index = finalArray.indexOf( forceRemoveSection.section.replace(/[0-9]/g, removing_index) );
+					var index = finalArray.indexOf( forceRemoveSection.section.replace(/[0-9]+/g, removing_index) );
 					if (index !== -1) {
 						finalArray.splice(index, 1);
 					}
@@ -81,6 +91,12 @@ const ItemComponent = props => {
 
 				{ choices[props.item]['delete'] &&  <span title="Delete"
 					  onClick={e => {
+
+						  // Skip clone if already is in progress.
+						  if( sessionStorage.getItem('forceRemoveInProgress') ) {
+							  return;
+						  }
+
 						  e.stopPropagation();
 						  deleteItem(props);
 						  props.removeItem(props.item);
